+++ index.d.ts
@@ -1,6 +1,0 @@
-/**
- * 3 Letter currency code
- *
- * _Accepts either uppercase or lowercase_
- */
-export default function(currencyCode: string): string