+++ changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+### 1.0.0
+
+##### 22/08/2021
+
+- Convert module to TypeScript
+- Updated development dependencies
+
+##### BREAKING CHANGE
+
+- Uses named export, default export is no longer supported
+
+```js
+import { getCurrencySymbol } from "@brixtol/currency-symbols";
+```
+
 ### 0.1.3
 
 ##### 19/10/2020